requirements.txt
python-telegram-bot==20.3
pandas
ta
requests
python-dotenv
telegram
flask
apscheduler
analyzer
